+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Search on the folders screen ignoring folder paths
+- Inability to focus the video timeline or the buttons using the "Down" button
+  on the media viewer screen
 
 ## [1.31.0] - 2024-09-28